added MyOmronTokenData as an internal model to contain Omron token data, and now storing this in the user's workspace so it's available with the rest of the user's session data, for issue #25

+++ src/main/java/edu/ohsu/cmp/coach/model/omron/AccessTokenResponse.java
@@ -3,7 +3,7 @@ package edu.ohsu.cmp.coach.model.omron;
 public class AccessTokenResponse {
     private String idToken;         // id of the user who performed the authorization
     private String accessToken;
-    private Long expiresIn;
+    private Integer expiresIn;
     private String tokenType;
     private String refreshToken;
 
@@ -23,11 +23,11 @@ public class AccessTokenResponse {
         this.accessToken = accessToken;
     }
 
-    public Long getExpiresIn() {
+    public Integer getExpiresIn() {
         return expiresIn;
     }
 
-    public void setExpiresIn(Long expiresIn) {
+    public void setExpiresIn(Integer expiresIn) {
         this.expiresIn = expiresIn;
     }
